Fix Build on UE 5.0.3

Former-commit-id: c14b8c6cf4a314ae92618f34af3e1ec9c1b2c945 [formerly a1c5910895c15376f36ab282f7052d707bac18b6]
Former-commit-id: 90c7e58d3efd60b3db2b33983e219a10605d3cc1

+++ Source/VoxelGraphEditor/Private/Preview/SVoxelGraphPreview.cpp
@@ -120,7 +120,7 @@ FReply SVoxelGraphPreview::OnMouseButtonDown(const FGeometry& MyGeometry, const 
 		const FVector2D RelativePosition = LocalClickPosition / Size;
 
 		const FVoxelGraphPreviewSettingsWrapper Wrapper(*PreviewSettings);
-		PreviewSettings->PreviewedVoxel = Wrapper.Start + FVoxelUtilities::RoundToInt(Wrapper.GetRelativePosition(RelativePosition.X, 1 - RelativePosition.Y) * FVector(Wrapper.Size * Wrapper.Step));
+		PreviewSettings->PreviewedVoxel = Wrapper.Start + FVoxelUtilities::RoundToInt(Wrapper.GetRelativePosition(static_cast<float>(RelativePosition.X), 1.f - RelativePosition.Y) * FVector(Wrapper.Size * Wrapper.Step));
 
 		if (!PreviewSettings->bShowStats && !PreviewSettings->bShowValues)
 		{
@@ -147,7 +147,7 @@ FReply SVoxelGraphPreview::OnMouseButtonUp(const FGeometry& MyGeometry, const FP
 		const FVector2D RelativePosition = Position / Size;
 		
 		const FVoxelGraphPreviewSettingsWrapper Wrapper(*PreviewSettings);
-		PreviewSettings->Center -= FVoxelUtilities::RoundToInt(Wrapper.GetRelativePosition(RelativePosition.X, -RelativePosition.Y) * FVector(Wrapper.Size * Wrapper.Step));
+		PreviewSettings->Center -= FVoxelUtilities::RoundToInt(Wrapper.GetRelativePosition(static_cast<float>(RelativePosition.X), static_cast<float>(- RelativePosition.Y)) * FVector(Wrapper.Size * Wrapper.Step));
 
 		FPropertyChangedEvent PropertyChangedEvent(UVoxelGraphPreviewSettings::StaticClass()->FindPropertyByName(GET_MEMBER_NAME_CHECKED(UVoxelGraphPreviewSettings, Center)));
 		PreviewSettings->PostEditChangeProperty(PropertyChangedEvent);
